requirements.txt
disnake
apscheduler
pillow
requests
dynamicprompts
aiohttp
temporalio
